agent/requirements.txt
web3
eth-account
requests
python-dotenv
openai
fastapi
uvicorn
pydantic